Extract review scorecard into modular feature slices
